+++ Nutrons2014/resources/META-INF/MANIFEST.MF
@@ -1,6 +1,6 @@
 MIDlet-Name: Nurtrons2014
 MIDlet-Version: 1.0.0
 MIDlet-Vendor: FIRST
-MIDlet-1: Nutrons2014, , edu.wpi.first.wpilibj.templates.Nutrons2014
+MIDlet-1: Nutrons2014, ,  com.nutrons.aerialassist.Nutrons2014
 MicroEdition-Profile: IMP-1.0
 MicroEdition-Configuration: CLDC-1.1
