+++ Nutrons2014/src/com/nutrons/aerialassist/Nutrons2014.java
@@ -5,15 +5,15 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.first.wpilibj.templates;
+package com.nutrons.aerialassist;
 
 
 import edu.wpi.first.wpilibj.IterativeRobot;
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
-import edu.wpi.first.wpilibj.templates.commands.CommandBase;
-import edu.wpi.first.wpilibj.templates.commands.drivetrain.DTManualTankCmd;
+import com.nutrons.aerialassist.commands.CommandBase;
+import com.nutrons.aerialassist.commands.drivetrain.DTManualTankCmd;
 
 /**
  * The VM is configured to automatically run this class, and to call the
