+++ Nutrons2014/src/com/nutrons/aerialassist/subsystems/ExampleSubsystem.java
@@ -1,5 +1,5 @@
 
-package edu.wpi.first.wpilibj.templates.subsystems;
+package com.nutrons.aerialassist.subsystems;
 
 import edu.wpi.first.wpilibj.command.Subsystem;
